Used SubscriberId as Id for NotificationCommand

+++ NetflixServer.Shared/NotificationCommand.cs
@@ -4,13 +4,11 @@ namespace NetflixServer.Shared
 {
     public class NotificationCommand
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         public string Email { get; set; }
 
         public string UserName { get; set; }
-
-        //public string Content { get; set; }
 
         public bool Active { get; set; }
